+++ src/Semantic/CLI.hs
@@ -17,6 +17,7 @@ import           Semantic.Api hiding (File)
 import qualified Semantic.AST as AST
 import           Semantic.Config
 import qualified Semantic.Graph as Graph
+import qualified Semantic.Parse as Parse
 import qualified Semantic.Task as Task
 import qualified Semantic.Git as Git
 import           Semantic.Task.Files
@@ -57,13 +58,13 @@ main = do
   (options, task) <- customExecParser (prefs showHelpOnEmpty) arguments
   config <- defaultConfig options
   res <- withTelemetry config $ \ (TelemetryQueues logger statter _) ->
-    Task.runTask (Task.TaskSession config "-" (optionsLogPathsOnError options) logger statter) task
+    Task.runTask (Task.TaskSession config "-" (optionsLogPathsOnError options) logger statter) (Parse.runParse task)
   either (die . displayException) pure res
 
 -- | A parser for the application's command-line arguments.
 --
 --   Returns a 'Task' to read the input, run the requested operation, and write the output to the specified output path or stdout.
-arguments :: ParserInfo (Options, Task.TaskC ())
+arguments :: ParserInfo (Options, Parse.ParseC Task.TaskC ())
 arguments = info (version <*> helper <*> ((,) <$> optionsParser <*> argumentsParser)) description
   where
     version = infoOption versionString (long "version" <> short 'v' <> help "Output the version of the program")
@@ -79,13 +80,13 @@ optionsParser = do
   logPathsOnError <- switch (long "log-paths" <> help "Log source paths on parse and assignment error.")
   pure $ Options logLevel logPathsOnError (Flag.flag FailOnWarning failOnWarning) (Flag.flag FailOnParseError failOnParseError)
 
-argumentsParser :: Parser (Task.TaskC ())
+argumentsParser :: Parser (Parse.ParseC Task.TaskC ())
 argumentsParser = do
   subparser <- hsubparser (diffCommand <> parseCommand <>  tsParseCommand <> graphCommand)
   output <- ToPath <$> strOption (long "output" <> short 'o' <> help "Output path, defaults to stdout") <|> pure (ToHandle stdout)
   pure $ subparser >>= Task.write output
 
-diffCommand :: Mod CommandFields (Task.TaskC Builder)
+diffCommand :: Mod CommandFields (Parse.ParseC Task.TaskC Builder)
 diffCommand = command "diff" (info diffArgumentsParser (progDesc "Compute changes between paths"))
   where
     diffArgumentsParser = do
@@ -98,7 +99,7 @@ diffCommand = command "diff" (info diffArgumentsParser (progDesc "Compute change
       filesOrStdin <- Right <$> some (Both <$> argument filePathReader (metavar "FILE_A") <*> argument filePathReader (metavar "FILE_B")) <|> pure (Left stdin)
       pure $ Task.readBlobPairs filesOrStdin >>= renderer
 
-parseCommand :: Mod CommandFields (Task.TaskC Builder)
+parseCommand :: Mod CommandFields (Parse.ParseC Task.TaskC Builder)
 parseCommand = command "parse" (info parseArgumentsParser (progDesc "Generate parse trees for path(s)"))
   where
     parseArgumentsParser = do
@@ -146,7 +147,7 @@ parseCommand = command "parse" (info parseArgumentsParser (progDesc "Generate pa
                   <|> pure (FilesFromHandle stdin)
       pure $ Task.readBlobs filesOrStdin >>= runReader languageModes . renderer
 
-tsParseCommand :: Mod CommandFields (Task.TaskC Builder)
+tsParseCommand :: Mod CommandFields (Parse.ParseC Task.TaskC Builder)
 tsParseCommand = command "ts-parse" (info tsParseArgumentsParser (progDesc "Generate raw tree-sitter parse trees for path(s)"))
   where
     tsParseArgumentsParser = do
@@ -165,7 +166,7 @@ tsParseCommand = command "ts-parse" (info tsParseArgumentsParser (progDesc "Gene
                   <|> pure (FilesFromHandle stdin)
       pure $ Task.readBlobs filesOrStdin >>= AST.runASTParse format
 
-graphCommand :: Mod CommandFields (Task.TaskC Builder)
+graphCommand :: Mod CommandFields (Parse.ParseC Task.TaskC Builder)
 graphCommand = command "graph" (info graphArgumentsParser (progDesc "Compute a graph for a directory or from a top-level entry point module"))
   where
     graphArgumentsParser = makeGraphTask
