+++ src/Semantic/REPL.hs
@@ -37,6 +37,7 @@ import Semantic.Analysis
 import Semantic.Config (logOptionsFromConfig)
 import Semantic.Distribute
 import Semantic.Graph
+import Semantic.Parse
 import Semantic.Resolution
 import Semantic.Task hiding (Error)
 import qualified Semantic.Task.Files as Files
