+++ src/Semantic/Util.hs
@@ -47,6 +47,7 @@ import           Prologue
 import           Semantic.Analysis
 import           Semantic.Config
 import           Semantic.Graph
+import           Semantic.Parse
 import           Semantic.Task
 import           Source.Loc
 import           System.Exit (die)
@@ -102,7 +103,7 @@ evaluateProject proxy parser paths = withOptions debugOptions $ \ config logger 
 
 -- Evaluate a project consisting of the listed paths.
 evaluateProject' session proxy parser paths = do
-  res <- runTask session $ do
+  res <- runTask session . runParse $ do
     blobs <- catMaybes <$> traverse readBlobFromFile (flip File (Language.reflect proxy) <$> paths)
     package <- fmap (quieterm . snd) <$> parsePackage parser (Project (takeDirectory (maybe "/" fst (uncons paths))) blobs (Language.reflect proxy) [])
     modules <- topologicalSort <$> runImportGraphToModules proxy package
@@ -120,9 +121,9 @@ parseFile, parseFileQuiet :: Parser term -> FilePath -> IO term
 parseFile parser = runTask' . (parse parser <=< readBlob . fileForPath)
 parseFileQuiet parser = runTaskQuiet . (parse parser <=< readBlob . fileForPath)
 
-runTask', runTaskQuiet :: TaskC a -> IO a
-runTask' task = runTaskWithOptions debugOptions task >>= either (die . displayException) pure
-runTaskQuiet task = runTaskWithOptions defaultOptions task >>= either (die . displayException) pure
+runTask', runTaskQuiet :: ParseC TaskC a -> IO a
+runTask' task = runTaskWithOptions debugOptions (runParse task) >>= either (die . displayException) pure
+runTaskQuiet task = runTaskWithOptions defaultOptions (runParse task) >>= either (die . displayException) pure
 
 mergeErrors :: Either (SomeError (Sum errs)) (Either (SomeError err) result) -> Either (SomeError (Sum (err ': errs))) result
 mergeErrors = either (\ (SomeError sum) -> Left (SomeError (weaken sum))) (either (\ (SomeError err) -> Left (SomeError (inject err))) Right)
